+++ CHANGELOG.md
@@ -26,7 +26,7 @@
   - Fix `CONSUMABLES_VIEW` off-by-one
   - Update position for buttons in `Enemies` room
   - Update `roomsDone` when needed
-  - Improved `OPEN_BAG_VIEW`
+  - Improved `OPEN_BAG_VIEW`, `EQUIPS_VIEW`
   - Proper centering of player in `draw_floor_view()`
 - Improve game loading in `raylib` build
   - Prep `Floor` on load, set `done_loading` in some cases